ci: disable netlify asset optimization
Update the netlify configuration file to directly include the parameters
previsouly set via Netlify's UI, removing the asset optimization
configuration along the way as this feature is deprecated and will be at
end of service on October 17, 2023.

This is actually one of the main reason why we moved to a more complexe
website build using rollup in the first place.

see https://answers.netlify.com/t/please-read-deprecation-of-post-processing-asset-optimization/96657